Doc: How to use with a GUI editor
Since the `%s` feature was removed, it is particularly difficult to find
the right command to launch an editor with a graphical user interface
(for instance, Neovim inside a terminal emulator).

This commit adds an example that users can use as-is or use as a base
for their custom command.